+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Granular datastore connection flags (`--datastore-host`, `--datastore-port`, `--datastore-username`, `--datastore-password`, `--datastore-database`, `--datastore-conn-param`) as an alternative to providing a full `--datastore-conn-uri` (https://github.com/authzed/spicedb/pull/2604)
 
 ## [1.56.1] - 2026-08-26
 ### Changed